+++ src/client/widgets/common/widget.js
@@ -216,7 +216,9 @@ class Widget extends EventEmitter {
                 event: 'value'
             }, builtIn: true, parent: this})
             this.on('change', (e)=>{
-                if (e.widget === this && this.mounted && !e.options.fromEdit) this.script.setValue(this.value, {...e.options, id: e.options.id || e.id})
+                if (e.widget === this && this.mounted && !e.options.fromEdit) {
+                    this.script.setValue(e.options.widget ? e.options.widget.value : this.value, {...e.options, id: e.options.id || e.id})
+                }
             })
             if (String(this.getProp('script')).includes('touch')) {
                 this.on('touch', (e)=>{
@@ -659,7 +661,7 @@ class Widget extends EventEmitter {
             } catch (err) {
                 let stackline = err.stack ? (err.stack.match(/>:([0-9]+):[0-9]+/) || '') : '',
                     line = stackline.length > 1 ? ' at line ' + (parseInt(stackline[1]) - 2) : ''
-                console.log((this.getProp('id') || this.props.id) + '.' + propName + ': JS{{}} error:\n' + err + line)
+                console.error((this.getProp('id') || this.props.id) + '.' + propName + ': JS{{}} error:\n' + err + line)
             }
 
             try {
@@ -680,7 +682,7 @@ class Widget extends EventEmitter {
             } catch (err) {
                 let stackline = err.stack ? (err.stack.match(/>:([0-9]+):[0-9]+/) || '') : '',
                     line = stackline.length > 1 ? ' at line ' + (parseInt(stackline[1]) - 2) : ''
-                console.log((this.getProp('id') || this.props.id) + '.' + propName + ': #{} error:\n' + err + line)
+                console.error((this.getProp('id') || this.props.id) + '.' + propName + ': #{} error:\n' + err + line)
             }
 
             for (let k in variables) {