+++ src/client/widgets/scripts/script-vm.js
@@ -59,7 +59,7 @@ class ScriptVm extends Vm {
 
         super.registerGlobals()
 
-        this.sandbox.contentWindow.set = (id, value)=>{
+        this.sandbox.contentWindow.set = (id, value, extraOptions = {send: true, sync: true})=>{
 
             var options = this.getValueOptions()
             options.fromScript = true
@@ -67,11 +67,24 @@ class ScriptVm extends Vm {
             // if (id === options.id) options.sync = false // loop stop
             // if (this.getWidget() === options.widget) options.sync = false // loop stop
 
-            var widgets = this.resolveId(id)
+            if (extraOptions.send === false) options.send = false
+            if (extraOptions.sync === false) options.sync = false
+
+            var widgets
+            if (id.includes('*')) {
+                var widget = this.getWidget()
+                if (widget.builtIn) widget = widget.parent
+                widgets = this.resolveId(
+                    Object.keys(widgetManager.idRoute).filter(key => key.match(new RegExp('^' + id.replace(/\*/g, '.*') + '$')))
+                ).filter(w => w !== widget)
+            } else {
+                widgets = this.resolveId(id).slice(0, 1)
+            }
+
 
             for (var i = widgets.length - 1; i >= 0; i--) {
 
-                return widgets[i].setValue(value, options)
+                widgets[i].setValue(value, options)
 
             }
 
@@ -100,7 +113,7 @@ class ScriptVm extends Vm {
             var widget = this.getWidget()
             if (widget.builtIn) widget = widget.parent
 
-            widget.sendValue(overrides)
+            widget.sendValue(overrides, {force: true})
 
         }
 
@@ -146,6 +159,13 @@ class ScriptVm extends Vm {
                 widgets[i].updateProps(Array.isArray(prop) ? prop : [prop], widget)
 
             }
+
+        }
+
+        this.sandbox.contentWindow.getIndex = (id = 'this')=>{
+
+            var widget = this.resolveId(id).pop()
+            if (widget) return widget.parent.children.indexOf(widget)
 
         }
 
@@ -319,7 +339,7 @@ class ScriptVm extends Vm {
 
             if (!options.send) return
 
-            toolbar = toolbar || require('../../ui/ui-toolbar')
+            toolbar = toolbar || require('../../ui/main-menu')
 
             var action = toolbar.entries.filter(x=>!x.separator)
 
@@ -333,7 +353,7 @@ class ScriptVm extends Vm {
 
         }
 
-        for (var imports of ['set', 'get', 'getProp', 'updateProp', 'send', 'httpGet', 'stateGet', 'stateSet', 'storage',
+        for (var imports of ['set', 'get', 'getProp', 'getIndex', 'updateProp', 'send', 'httpGet', 'stateGet', 'stateSet', 'storage',
             'setInterval', 'clearInterval', 'setTimeout', 'clearTimeout', 'unfocus', 'setScroll', 'getScroll', 'toolbar']) {
             this.sanitize(this.sandbox.contentWindow[imports])
         }